requirements.txt
Flask==3.0.3
flask-cors==4.0.1
flask-openapi3[swagger]==3.1.0
Flask-SQLAlchemy==3.1.1
pydantic>=2.0
